fix error from comment
