+++ app/utils/html2canvas-clone.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/utils/html2canvas-clone';
